Add support to clean download text files (#7)
It seems bitar does not handle text files well when downloading an
archive and reconciling chunks in an existing file. Everytime we update
an XML file the updater corrupts the existing one on disk and users need
to delete the file and updater manifest to force a clean download. This
commit adds a special handling case for text files where it will always
delete the on disk file first forcing bitar to do a full clean download
of the remote archive.

Additionally, some clippy warnings were fixed and a bug where the play
button would break when running the updater from a script was fixed.

Closes #6