Rewrite updater and port to egui
The core of the updater was re-written and simplified. More work is done
concurrently so it is faster and the logic is now much simpler to follow
and reason about. Additionally progress and state is better reported
with distinctions made between actual downloading and verifying the
files.

The user interface was completely rewritten to remove FLTK and the
webview in favor of egui. News is now pulled from the website using the
API rather than showing the image tiles. Additionally a new options view
was added to allow users to re-verify the files (i.e. no more deleting
updater folder), find their configuration folder or open the support
website.

Finally, a temporary checkbox was added to support launching the beta
client that is soon to be released.