Ralph/egui (#21)
Fixed compile errors when attempting to compile.
Added some arguments for auto launching the game when the update is
complete.

I had originally made big changes for the updater to be able to run as a
native Linux app but decided against it as it would require many changes
that I wasn't happy with.

Also with the final goal of being released on Steam and being able to
run on the Steam deck, those changes wouldn't make much sense.

I have some improvements that I do want to make but it will require game
client changes. Mainly I would like to make it so we can skip the login
section of the game and just jump right into character selection.

Once again, I have tested and confirmed that this version does work on
linux under wine.